# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School API",
    "version": "1.0.0",
    "description": "API for managing schools"
  },
  "paths": {
    "/api/addSchool": {
      "post": {
        "summary": "Add a new school",
        "description": "Creates a new school with the provided data",
        "operationId": "addSchool",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the school"
                  },
                  "address": {
                    "type": "string",
                    "description": "The address of the school"
                  },
                  "latitude": {
                    "type": "number",
                    "description": "Latitude of the school"
                  },
                  "longitude": {
                    "type": "number",
                    "description": "Longitude of the school"
                  }
                },
                "required": ["name", "address", "latitude", "longitude"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "School successfully added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "latitude": {
                          "type": "number"
                        },
                        "longitude": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/listSchools": {
      "get": {
        "summary": "List schools sorted by proximity",
        "description": "Fetches all schools and sorts them by proximity to the provided coordinates",
        "operationId": "listSchools",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude of the user location"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude of the user location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of schools sorted by proximity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "latitude": {
                            "type": "number"
                          },
                          "longitude": {
                            "type": "number"
                          },
                          "distance": {
                            "type": "number",
                            "description": "Distance from the provided coordinates in km"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Latitude and longitude are required and must be valid numbers"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/getSchool": {
      "get": {
        "summary": "Get a specific school by ID",
        "description": "Fetches a specific school based on the provided ID",
        "operationId": "getSchool",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID of the school to fetch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "latitude": {
                          "type": "number"
                        },
                        "longitude": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/updateSchool/{id}": {
      "put": {
        "summary": "Update a school",
        "description": "Updates the details of a specific school by ID",
        "operationId": "updateSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID of the school to update"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "School successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "latitude": {
                          "type": "number"
                        },
                        "longitude": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "School not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/deleteSchool/{id}": {
      "delete": {
        "summary": "Delete a school",
        "description": "Deletes a specific school by ID",
        "operationId": "deleteSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID of the school to delete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}